requirements.txt
requests~=2.32.4
fastapi~=0.115.13
pydantic~=2.11.7
pdfplumber~=0.11.7
langchain~=0.3.27
langchain-community~=0.3.27
groq~=0.29.0
python-dotenv~=1.1.0
uvicorn
chromadb
sentence-transformers